feat: error handler

+++ api/window/main.py
@@ -1,12 +1,16 @@
-from hashlib import new
 import os
+import os.path
 
 class Main():
     def main():
         old_name = "./dist/update.exe" 
         new_name = "./dist/Update Data.exe" 
 
-        os.rename(old_name, new_name)
+        if(os.path.exists(new_name)):
+            os.remove(new_name)
+
+        if(os.path.exists(old_name)):
+            os.rename(old_name, new_name)
 
 
 if __name__ == "__main__":
